// src/types/riot/AccountTier.ts
/**
 * Tiers that an account can be.
 * @public
 */
enum AccountTier {
	/** Iron. */
	IRON = "IRON",

	/** Bronze. */
	BRONZE = "BRONZE",

	/** Silver. */
	SILVER = "SILVER",

	/** Gold. */
	GOLD = "GOLD",

	/** Platinum. */
	PLATINUM = "PLATINUM",

	/** Emerald. */
	EMERALD = "EMERALD",

	/** Diamond. */
	DIAMOND = "DIAMOND",

	/** Master. */
	MASTER = "MASTER",

	/** Grandmaster. */
	GRANDMASTER = "GRANDMASTER",

	/** Challenger. */
	CHALLENGER = "CHALLENGER"
}

export default AccountTier;